feat: Cleanup the Error story; remove SecretKey::as_bytes
